+++ guillotina/tests/conftest.py
@@ -1,9 +1,6 @@
 from pytest_docker_fixtures import images
 
 
-images.configure("cockroach", "cockroachdb/cockroach", "v2.1.6")
-
-
 images.configure("postgresql", version="10.9")