+++ guillotina/tests/test_catalog.py
@@ -28,7 +28,7 @@ import pytest
 pytestmark = pytest.mark.asyncio
 
 
-NOT_POSTGRES = os.environ.get("DATABASE", "DUMMY") in ("cockroachdb", "DUMMY")
+NOT_POSTGRES = os.environ.get("DATABASE", "DUMMY") in ("DUMMY")
 PG_CATALOG_SETTINGS = {
     "applications": ["guillotina.contrib.catalog.pg"],
     "load_utilities": {