+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ CHANGELOG
 6.0.9 (unreleased)
 ------------------
 
-- Fix bug on error deserialization [lferran]
+- Fix bug on error deserialization
 
 - Fix transaction context manager doesn't abort the txn when a exception is raised
   [masipcat]